requirements.txt
# Core dependencies
numpy
opencv-python
pandas
matplotlib
scipy
tqdm
Pillow
pyvista